add device.cpp for splitting device class
